+++ HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ---------------
 
+2.13.6 (2026-03-17)
++++++++++++++++++++
+
+**Improvements**
+
+- live_orders threading Event added to Markets
+- complete orders logic executed on each market per update
+
 2.13.5 (2026-03-16)
 +++++++++++++++++++